+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.36.2] - 2026-04-02
+
 ### Fixed
 
 - upcloud_file_storage_share: fix resource import by parsing the file storage service uuid and share name from id
@@ -1204,7 +1206,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.36.1...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.36.2...HEAD
+[5.36.2]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.36.1...v5.36.2
 [5.36.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.36.0...v5.36.1
 [5.36.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.35.0...v5.36.0
 [5.35.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.34.1...v5.35.0